eval speed simd
Score of Koivisto Dev vs Koivisto Test: 328 - 271 - 525 [0.525]
Elo difference: 17.63 +/- 14.82

1124 of 5000 games finished.

Former-commit-id: 291f57db7e560bbfc40be679f3ab6084483d5647

+++ src_files/eval.h
@@ -19,24 +19,24 @@ class Evaluator{
          * returns the phase of the last calculation
          * @return
          */
-        double getPhase();
+        float getPhase();
         
         /**
          * returns a list of features of the last calculation
          * @return
          */
-        double *getFeatures();
+        float *getFeatures();
         
         /**
          * returns a list of early game parameters
          * @return
          */
-        double *getEarlyGameParams();
+        float *getEarlyGameParams();
         
         /**
          * returns a list of late game parameters
          */
-        double *getLateGameParams();
+        float *getLateGameParams();
     
         /**
          * returns the amount of tunable parameters
